WIP: BIOSYS-308 Add API GET /Record end-point a boundary parameter
